pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cerevyn-agent"
version = "0.1.0"
description = "Autonomous multi-step AI agent (PS3)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.2.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
extraPaths = ["src"]
